moved database_connection module
