+++ spatialprofilingtoolbox/db/database_connection.py
@@ -5,7 +5,7 @@ import configparser
 
 import psycopg2
 
-from .log_formats import colorized_logger
+from ..log_formats import colorized_logger
 logger = colorized_logger(__name__)
